requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.22
mysql-connector-python==8.0.33
redis==4.6.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
openai==1.2.0
numpy==1.26.0
opencv-python==4.8.1.78
mediapipe==0.10.7
ffmpeg-python==0.2.0
reportlab==4.0.7
openpyxl==3.1.2
google-cloud-speech==2.21.0
whisper==1.1.10
aiofiles==23.2.1
email-validator==2.0.0